+++ src/main/java/com/epam/reportportal/service/step/StepReporter.java
@@ -24,6 +24,7 @@ import io.reactivex.Maybe;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.io.File;
+import java.util.Calendar;
 import java.util.Date;
 import java.util.function.Supplier;
 
@@ -118,25 +119,37 @@ public interface StepReporter {
 	@Nullable
 	<T> T step(@Nonnull String name, @Nonnull Supplier<T> actions);
 
+	default void setStepStatus(@Nonnull ItemStatus status) {
+	}
+
 	class StepEntry {
 		private final Maybe<String> itemId;
 		private final Date timestamp;
 		private final FinishTestItemRQ finishTestItemRQ;
 
-		public StepEntry(Maybe<String> itemId, Date timestamp, FinishTestItemRQ finishTestItemRQ) {
+		public StepEntry(@Nonnull Maybe<String> itemId, @Nonnull Date stepStartTime,
+		                 @Nonnull FinishTestItemRQ finishTestItemRQ) {
 			this.itemId = itemId;
-			this.timestamp = timestamp;
+			this.timestamp = stepStartTime;
 			this.finishTestItemRQ = finishTestItemRQ;
 		}
 
+		public StepEntry(@Nonnull Maybe<String> itemId,
+				@Nonnull FinishTestItemRQ finishTestItemRQ) {
+			this(itemId, Calendar.getInstance().getTime(), finishTestItemRQ);
+		}
+
+		@Nonnull
 		public Maybe<String> getItemId() {
 			return itemId;
 		}
 
+		@Nonnull
 		public Date getTimestamp() {
 			return timestamp;
 		}
 
+		@Nonnull
 		public FinishTestItemRQ getFinishTestItemRQ() {
 			return finishTestItemRQ;
 		}