+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- AspectJ version updated on 1.9.19 to support newer versions of Java and bug fixes, by @HardNorth
+- Some classes were refactored to less use Guava library and more core Java, by @HardNorth
+### Added
+- `ClientIdProvider` class to store and read a client's unique ID, by @HardNorth
+- `StepReporter.setStepStatus` method to change nested step status in runtime, by @HardNorth
 
 ## [5.1.18]
 ### Added